wed even

+++ src/java/AssayRunViewer.java
@@ -75,7 +75,7 @@ public class AssayRunViewer extends JDialog implements java.awt.event.ActionList
     //    this.session = dmf.getSession();
      IFn getProjectID = Clojure.var("ln.codax-manager", "get-project-id");
 
-     project_id = (int)getProjectID.invoke();
+     project_id = ((Integer)getProjectID.invoke()).intValue();
       IFn getUser = Clojure.var("ln.codax-manager", "get-user");
    
     owner = (String)getUser.invoke();
@@ -88,7 +88,7 @@ public class AssayRunViewer extends JDialog implements java.awt.event.ActionList
     assay_runs_pane_border.setTitlePosition(javax.swing.border.TitledBorder.TOP);
     assay_runs_pane.setBorder(assay_runs_pane_border);
 
-    assay_runs_table = dbm.getDatabaseRetriever().getAssayRuns((int)getProjectID.invoke());
+    assay_runs_table = dbm.getDatabaseRetriever().getAssayRuns(((Long)getProjectID.invoke()).intValue());
   assay_runs_table.getSelectionModel().addListSelectionListener(						     
 	  new ListSelectionListener() {
 	      public void valueChanged(ListSelectionEvent e) {
@@ -146,7 +146,7 @@ public class AssayRunViewer extends JDialog implements java.awt.event.ActionList
     hit_lists_pane_border.setTitlePosition(javax.swing.border.TitledBorder.TOP);
     hit_lists_pane.setBorder(hit_lists_pane_border);
 
- hit_lists_table = dbm.getDatabaseRetriever().getHitLists((int)getProjectID.invoke());
+    hit_lists_table = dbm.getDatabaseRetriever().getHitLists(((Long)getProjectID.invoke()).intValue());
 
     hit_lists_scroll_pane = new JScrollPane(hit_lists_table);
     hit_lists_table.setFillsViewportHeight(true);
