+++ src/java/HitListViewer.java
@@ -68,7 +68,7 @@ public class HitListViewer extends JDialog implements java.awt.event.ActionListe
 	this.setTitle("Hit List Viewer - " + (String)getProjectSysName.invoke());
 	IFn getProjectID = Clojure.var("ln.codax-manager", "get-project-id");
 
-	project_id = (int)getProjectID.invoke();
+	project_id = ((Long)getProjectID.invoke()).intValue();
       IFn getUser = Clojure.var("ln.codax-manager", "get-user");
 
       owner = (String)getUser.invoke();
@@ -94,7 +94,7 @@ public class HitListViewer extends JDialog implements java.awt.event.ActionListe
       JPanel     hits_buttons = new JPanel(buttonLayout);
       
       //get all the hit lists in the current project       
-      current_project_id = (int)getProjectID.invoke();
+      current_project_id = ((Long)getProjectID.invoke()).intValue();
       all_hit_lists_in_project = new JComboBox<ComboItem>(dbm.getDatabaseRetriever().getHitListsForProject(current_project_id));
     for(int i=0; i < all_hit_lists_in_project.getItemCount(); i++){
 	if(((ComboItem)all_hit_lists_in_project.getItemAt(i)).getKey() == current_project_id){
@@ -136,7 +136,7 @@ public class HitListViewer extends JDialog implements java.awt.event.ActionListe
     counts_paneBorder.setTitlePosition(javax.swing.border.TitledBorder.TOP);
     counts_pane.setBorder(counts_paneBorder);
 
-    counts_table = dbm.getDatabaseRetriever().getHitCountPerPlateSet((int)getProjectID.invoke(), hit_list_id);
+    counts_table = dbm.getDatabaseRetriever().getHitCountPerPlateSet(((Long)getProjectID.invoke()).intValue(), hit_list_id);
 
     counts_scroll_pane = new JScrollPane(counts_table);
     counts_table.setFillsViewportHeight(true);
@@ -182,7 +182,7 @@ public class HitListViewer extends JDialog implements java.awt.event.ActionListe
 	hits_table.setModel(new_model); 
 	IFn getProjectID = Clojure.var("ln.codax-manager", "get-project-id");
  
-	JTable new_counts_table = dbm.getDatabaseRetriever().getHitCountPerPlateSet((int)getProjectID.invoke(), selected_hit_list_id);
+	JTable new_counts_table = dbm.getDatabaseRetriever().getHitCountPerPlateSet(((Long)getProjectID.invoke()).intValue(), selected_hit_list_id);
 	TableModel new_model2 = new_counts_table.getModel();
 	counts_table.setModel(new_model2);
